4.21: only authenticated creator can delete blog

+++ controllers/blogs.js
@@ -49,6 +49,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const toDelete = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (toDelete.user.toString() !== decodedToken.id.toString()) {
+    let err = new Error('wrong user')
+    err.name = 'AuthenticationError'
+    throw err
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
